refactor: moved doc structures from flux to fluxcore
