+++ libflux/flux/src/bin/fluxdoc.rs
@@ -1,7 +1,8 @@
 #![cfg_attr(feature = "strict", deny(warnings))]
 
 //use std::collections::HashMap;
-use flux::docs::{self, DocPackage};
+use fluxcore::semantic::bootstrap::DocPackage;
+use flux::docs::{self};
 use std::fs::{self, File};
 use std::io::Write;
 use std::path::{Path, PathBuf};
